gcs_transcript_retriever/requirements.txt
# requirements.txt
google-cloud-storage==2.14.0